Metadata is now state instead of ref
Changelist:
- metadata now is state of ref, allowing it to change while the application is
running. Both the bar information and the bpm can change while the player is
running, and the changes will take effect immediately.
- Changes the data map to use Beats as keys instead of ticks. This allows the
drumkit to "remember" the value of each beat when changing the loop information.
For example, ([bar 1 / beat 2 / subdivision 2], 1) will be used instead of
storing ([tick 5], 1), which means that changing the loop structure (change any
of bar/beat/subdivision) will cause the relative position of ticks to change,
but not the absolute positions.
- Changes the DrumkitXXX components to work with the new Map type.
- Changes Beat constructor to cache Beat instances to allow for the above map
to function, as JavaScript Maps use the SameValueZero algorithm to determine
whether two keys are the same.
- Updates the setData to actually cause a re-render, meaning that it is now
possible to programmatically change the data and subsequently the UI.
- Changes the DrumkitXXX components to propagate that state down to the
DrumkitCell.
- Updates documentation and cleans up code.